clippy and moves
